chore(codex-memory): share capture sanitization
