+++ examples/codex-memory-plugin/scripts/pre-compact-capture.mjs
@@ -21,6 +21,7 @@
  */
 
 import { readFile } from "node:fs/promises";
+import { sanitizeCapturedText } from "./captured-text.mjs";
 import { loadConfig } from "./config.mjs";
 import { describeCommitOutcome, waitForCommitTask } from "./commit-task.mjs";
 import { createLogger } from "./debug-log.mjs";
@@ -83,20 +84,6 @@ function parseTranscript(content) {
   for (const line of lines) {
     try { out.push(JSON.parse(line)); } catch { /* skip */ }
   }
-  return out;
-}
-
-function sanitizeCapturedText(text) {
-  let out = String(text || "");
-  if (/^\s*You are running as a subagent\b/i.test(out)) return "";
-  out = out
-    .replace(/<relevant-memories>[\s\S]*?<\/relevant-memories>/gi, "")
-    .replace(/<<<BEGIN_OPENCLAW_INTERNAL_CONTEXT>>>[\s\S]*?<<<END_OPENCLAW_INTERNAL_CONTEXT>>>/g, "")
-    .replace(/Conversation context \(untrusted metadata\):\s*```json[\s\S]*?```\s*/gi, "")
-    .replace(/\[Inter-session message][\s\S]*?(?=\n\[[a-z]+]:|\n?$)/gi, "")
-    .replace(/Full hook output saved to:\s*\S+/gi, "")
-    .trim();
-  if (/^(?:\[user]:\s*)*$/i.test(out)) return "";
   return out;
 }
